add high level test covering transactional id auth
Extracts a base AbstractTracingIT
